% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphStatistics.R
\name{cosineSimilarity}
\alias{cosineSimilarity}
\title{Compute cosine similarity coefficient of two networks}
\usage{
cosineSimilarity(i, j, type = "vertex")
}
\arguments{
\item{i}{igraph object}

\item{j}{igraph object}

\item{type}{character value - "vertex" or"edge", indicating structures to be compared}
}
\description{
Compute cosine similarity coefficient of two networks
}

