% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ga.R
\name{ga}
\alias{ga}
\title{Simple implementation of a binary Genetic Algorithm}
\usage{
ga(fitness, ..., pop.size = 50, n.bits, max.iter = 20, run = 10,
  pmutation = 0.01, eletism = 2, seed = 123)
}
\description{
Simple implementation of a binary Genetic Algorithm
}

