% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Genetic.R
\name{Genetic}
\alias{Genetic}
\title{Use genetic algorithm to identify maximally scoring subgraph}
\usage{
Genetic(graph, weights, pop.size = 75, max.iter = 1000, run = 20,
  p.mutation = 0.01, eletism = 1)
}
\arguments{
\item{graph}{igraph object}

\item{weights}{numeric vector}

\item{pop.size}{integer}

\item{max.iter}{integer}

\item{run}{integer}

\item{p.mutation}{float. Probability of mutation}

\item{eletism}{integer}
}
\description{
Use genetic algorithm to identify maximally scoring subgraph
}

