% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggnet.R
\name{ggnet}
\alias{ggnet}
\title{Plot a network using ggplot2}
\usage{
ggnet(g, node.lbl = NA, node.size = NA, node.col = NA, node.alpha = 0.5,
  edge.width, edge.col = "#3182bd")
}
\arguments{
\item{g}{igraph object}

\item{node.lbl}{character vector of graph node labels}

\item{node.size}{vector of integers proportional to the size of the node}

\item{node.col}{vector of intergers proportional to the colour of the node}

\item{node.alpha}{integer [0, 1] corresponding to node transparancy}

\item{edge.col}{character value representing edge colour}
}
\description{
Plot a network using ggplot2
}

