% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphStatistics.R
\name{GraphRecall}
\alias{GraphRecall}
\title{Compute network recall}
\usage{
GraphRecall(predicted, actual)
}
\arguments{
\item{predicted}{igraph object}

\item{actual}{igraph object}
}
\description{
Compute network recall
}

