% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Subgraph.R
\name{Subgraph}
\alias{Subgraph}
\title{Use genetic algorithm to identify maximally scoring subgraph}
\usage{
Subgraph(graph, weights, pop.size = 50, max.iter = 100,
  run = vcount(graph), p.mutation = 0.1, eletism = max(1, round(pop.size *
  0.05)), ncores)
}
\arguments{
\item{graph}{a named igraph object}

\item{weights}{a named numeric vector}

\item{pop.size}{integer}

\item{max.iter}{integer}

\item{run}{integer}

\item{p.mutation}{float. Probability of mutation}

\item{eletism}{integer}

\item{ncores}{integer}
}
\description{
Use genetic algorithm to identify maximally scoring subgraph
}

