% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gasub.R
\name{gaSubgraph}
\alias{gaSubgraph}
\title{Extract an active subnetwork using Genetic Algorithm scoring function optimisation}
\usage{
gaSubgraph(graph, v, optim = "weight", run = 50, maxiter = 1e+05,
  parallel = 2, ...)
}
\arguments{
\item{graph}{input igraph object}

\item{v}{named vector of node weights. Vector names must match igraph node
NAME attribute. In an input graph, nodes not found in V will be
assigned a weight of 0.}

\item{run}{the number of consecutive generations without any improvement
in the best finess value before the GA is stopped}

\item{maxiter}{the maximum number of iterations to run before the GA search
is halted.}

\item{parallel}{logical argument specifying if parallel computing should be
used (‘TRUE’) or not (‘FALSE’, default) for evaluating the
fitness function. This argument could also be used to specify
the number of cores to employ; by default, this is taken from
‘detectCores’. Finally, the functionality of parallelization
depends on system OS: on Windows only 'snow' type
functionality is available, while on Unix/Linux/Mac OSX both
'snow' and 'multicore' (default) functionalities are
available.}

\item{...}{additional arguments to be passed to GA:ga(function)}
}
\description{
Extract an active subnetwork using Genetic Algorithm scoring function optimisation
}

