% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gasub.R
\name{gaSubgraph}
\alias{gaSubgraph}
\title{Extract an active subnetwork using Genetic Algorithm scoring function optimisation}
\usage{
gaSubgraph(graph, weights, run = 50, max.iter = 1e+05)
}
\arguments{
\item{graph}{input igraph object}

\item{weights}{named vector of node weights. Vector names must match igraph node
NAME attribute. In an input graph, nodes not found in V will be
assigned a weight of 0.}

\item{run}{the number of consecutive generations without any improvement
in the best finess value before the GA is stopped}

\item{max.iter}{the maximum number of iterations to run before the GA search
is halted.}
}
\description{
Extract an active subnetwork using Genetic Algorithm scoring function optimisation
}

