% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Genetic.R
\name{Genetic}
\alias{Genetic}
\title{Use genetic algorithm to identify maximally scoring subgraph}
\usage{
Genetic(graph, weights, pop.size = 50, max.iter = 100, run = 100,
  p.mutation = 0.1, eletism = max(1, round(pop.size * 0.05)))
}
\arguments{
\item{graph}{igraph object}

\item{weights}{numeric vector}

\item{pop.size}{integer}

\item{max.iter}{integer}

\item{run}{integer}

\item{p.mutation}{float. Probability of mutation}

\item{eletism}{integer}
}
\description{
Use genetic algorithm to identify maximally scoring subgraph
}

