% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggnet.R
\name{ggnet}
\alias{ggnet}
\title{Plot a network using ggplot2}
\usage{
ggnet(g, node.size = NA, node.col = NA, node.alpha = 0.5,
  low.col = "blue", high.col = "red", edge.width, edge.col)
}
\arguments{
\item{g}{igraph object}

\item{node.size}{vector of integers proportional to the size of the node}

\item{node.col}{vector of intergers proportional to the colour of the node}

\item{node.alpha}{integer [0, 1] corresponding to node transparancy}

\item{low.col}{colour of the lowest value}

\item{high.col}{colour of the highest value}
}
\description{
Plot a network using ggplot2
}

