% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphStatistics.R
\name{GraphPrecision}
\alias{GraphPrecision}
\title{Compute network precision}
\usage{
GraphPrecision(predicted, actual)
}
\arguments{
\item{predicted}{igraph object}

\item{actual}{igraph object}
}
\description{
Compute network precision
}

