% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphStatistics.R
\name{JaccardSimilarity}
\alias{JaccardSimilarity}
\title{Compute jaccard similarity coefficient of two networks}
\usage{
JaccardSimilarity(i, j, type = "vertex")
}
\arguments{
\item{i}{igraph object}

\item{j}{igraph object}

\item{type}{character value - "vertex" or"edge", indicating structures to be compared}
}
\description{
Compute jaccard similarity coefficient of two networks
}

